netkvm: indicate network change on SRIOV attach
If the adapter does not have valid IP address yet, the
DHCP handshake will be triggered. Typical flow is:
The VIOPROT protocol is not installed
Disable VF adapter and NETKVM adapter
Enable VF adapter then enable NETKVM adapter
NETKVM does not receive IP address as the address for this
MAC is already assigned to VF.
Install the protocol. The VF does not use the IP address
anymore but the NETKVM can receive it only if DHCP timer
trigger the handshake. Now due to network change indication
the DHCP handshake is triggered immediately.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>