netkvm: process NetPnp events and detect protocol bind completion
The best event to detect protocol bind completion is
NetEventBindsComplete. I comes usually without parameters as it
indicates that all the suitable adapters bound to the protocol.
So we propagate it to all existing protocol bindings.
When we detect completion of protocol binding, we schedule work
item to attach the VF adapter to the NETKVM instance (if present).

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>